requirements.txt
flask
flask_sqlalchemy
werkzeug
pillow
torch
torchvision
gunicorn